AP-2626 Move Lead Proc Services to LF namespace
Moves the LeadApplicationProceedingTypeAssignmentService and
LeadProceedingAssignmentService into the LegalFramework namespace as
they seem to make more sense there.
